update (dashboard for a teacher)

+++ teachers/views.py
@@ -16,6 +16,8 @@ from teachers.models import Teacher
 # Create your views here.
 @login_required
 def teachers_list(request):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     teachers = Teacher.objects.annotate(
     num_students=Count('course__student', distinct=True),
     num_courses=Count("course", distinct=True)
@@ -27,6 +29,8 @@ def teachers_list(request):
 
 @login_required
 def add_teacher(request):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     if request.method == "POST":
         user = User.objects.create_user(username=request.POST['username'], password=request.POST['password'])
         Teacher.objects.create(
@@ -42,6 +46,8 @@ def add_teacher(request):
 
 @login_required
 def edit_teacher(request, id):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     if request.method == "POST":
         teacher = Teacher.objects.get(pk=id, center=request.user)
         teacher.first_name = request.POST.get('first_name')
@@ -57,6 +63,8 @@ def edit_teacher(request, id):
 
 @login_required
 def delete_confirmation_teacher(request, id):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     teacher = Teacher.objects.get(pk=id, center=request.user)
     data = {
         "teacher": teacher,
@@ -65,11 +73,15 @@ def delete_confirmation_teacher(request, id):
 
 @login_required
 def delete_teacher(request, id):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     Teacher.objects.get(pk=id, center=request.user).delete()
     return redirect('teachers-list')
 
 @login_required
 def teacher_details(request, id):
+    if hasattr(request.user, 'teacher_user'):
+        return redirect('teacher-dashboard')
     teacher = Teacher.objects.get(pk=id, center=request.user)
     courses = Course.objects.filter(course_teacher=teacher, center=request.user)
     students = Student.objects.annotate(
